+++ components/app/board/list/index.js
@@ -1,2 +1,1 @@
 export { Search } from './Search';
-export { ListItem } from './ListItem';
