+++ components/common/index.ts
@@ -8,3 +8,5 @@ export { Top5TextLoop } from './Top5TextLoop';
 export { UrlInput } from './UrlInput';
 export { TitleHeader } from './TitleHeader';
 export { MessageModal } from './MessageModal';
+export { ListAndPagination } from './ListandPagination';
+export { ListItem } from './ListItem';
